+++ firebase-dataconnect/src/androidTest/kotlin/com/google/firebase/dataconnect/testutil/schemas/RealtimeConnector.kt
@@ -18,6 +18,7 @@ package com.google.firebase.dataconnect.testutil.schemas
 
 import com.google.firebase.dataconnect.ConnectorConfig
 import com.google.firebase.dataconnect.FirebaseDataConnect
+import com.google.firebase.dataconnect.core.DataConnectGrpcRPCs
 import com.google.firebase.dataconnect.core.FirebaseDataConnectInternal
 import com.google.firebase.dataconnect.serializers.UUIDSerializer
 import com.google.firebase.dataconnect.testutil.DataConnectBackend
@@ -29,6 +30,9 @@ import kotlinx.serialization.serializer
 class RealtimeConnector private constructor(dataConnectInternal: FirebaseDataConnectInternal) {
 
   val dataConnect: FirebaseDataConnect = dataConnectInternal
+
+  internal val dataConnectGrpcRPCs: DataConnectGrpcRPCs by
+    lazy(LazyThreadSafetyMode.PUBLICATION) { dataConnectInternal.grpcRPCs }
 
   val resourceName: String = dataConnectInternal.connectorResourceName
 
@@ -57,7 +61,10 @@ class RealtimeConnector private constructor(dataConnectInternal: FirebaseDataCon
     fun queryRef(variables: Variables) =
       connector.dataConnect.query(OPERATION_NAME, variables, serializer<Data>(), serializer())
 
-    @Serializable data class Variables(val key: Key)
+    @Serializable
+    data class Variables(val key: Key) {
+      constructor(id: UUID) : this(Key(id))
+    }
 
     @Serializable
     data class Data(val item: Item?) {