+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/core/DataConnectGrpcRPCs.kt
@@ -30,6 +30,7 @@ import com.google.firebase.dataconnect.core.LoggerGlobals.debug
 import com.google.firebase.dataconnect.core.LoggerGlobals.warn
 import com.google.firebase.dataconnect.sqlite.DataConnectCacheDatabase
 import com.google.firebase.dataconnect.sqlite.GetEntityIdForPathFunction
+import com.google.firebase.dataconnect.util.CoroutineUtils
 import com.google.firebase.dataconnect.util.ImmutableByteArray
 import com.google.firebase.dataconnect.util.NullableReference
 import com.google.firebase.dataconnect.util.ProtoUtil.buildStructProto
@@ -42,6 +43,8 @@ import com.google.protobuf.Duration as DurationProto
 import com.google.protobuf.Struct
 import google.firebase.dataconnect.proto.ConnectorServiceGrpc
 import google.firebase.dataconnect.proto.ConnectorServiceGrpcKt
+import google.firebase.dataconnect.proto.ConnectorStreamServiceGrpc
+import google.firebase.dataconnect.proto.ConnectorStreamServiceGrpcKt
 import google.firebase.dataconnect.proto.EmulatorInfo
 import google.firebase.dataconnect.proto.EmulatorIssuesResponse
 import google.firebase.dataconnect.proto.EmulatorServiceGrpc
@@ -53,6 +56,8 @@ import google.firebase.dataconnect.proto.ExecuteQueryResponse
 import google.firebase.dataconnect.proto.GetEmulatorInfoRequest
 import google.firebase.dataconnect.proto.GraphqlResponseExtensions.DataConnectProperties
 import google.firebase.dataconnect.proto.StreamEmulatorIssuesRequest
+import google.firebase.dataconnect.proto.StreamRequest
+import google.firebase.dataconnect.proto.StreamResponse
 import io.grpc.ManagedChannel
 import io.grpc.ManagedChannelBuilder
 import io.grpc.Metadata
@@ -66,10 +71,15 @@ import kotlin.time.Duration
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.asExecutor
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.consumeAsFlow
 import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
+import kotlinx.coroutines.job
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withContext
@@ -79,6 +89,7 @@ internal class DataConnectGrpcRPCs(
   host: String,
   sslEnabled: Boolean,
   @get:VisibleForTesting val connectorResourceName: String,
+  nonBlockingCoroutineDispatcher: CoroutineDispatcher,
   private val blockingCoroutineDispatcher: CoroutineDispatcher,
   private val grpcMetadata: DataConnectGrpcMetadata,
   private val cacheSettings: CacheSettings?,
@@ -154,11 +165,24 @@ internal class DataConnectGrpcRPCs(
       ConnectorServiceGrpcKt.ConnectorServiceCoroutineStub(lazyGrpcChannel.getLocked())
     }
 
+  private val lazyStreamingGrpcStub =
+    SuspendingLazy(mutex) {
+      check(!closed) { "DataConnectGrpcRPCs ${logger.nameWithId} instance has been closed" }
+      ConnectorStreamServiceGrpcKt.ConnectorStreamServiceCoroutineStub(lazyGrpcChannel.getLocked())
+    }
+
   private val lazyEmulatorGrpcStub =
     SuspendingLazy(mutex) {
       check(!closed) { "DataConnectGrpcRPCs ${logger.nameWithId} instance has been closed" }
       EmulatorServiceGrpcKt.EmulatorServiceCoroutineStub(lazyGrpcChannel.getLocked())
     }
+
+  private val connectCoroutineScope =
+    CoroutineUtils.createSupervisorCoroutineScope(
+      nonBlockingCoroutineDispatcher,
+      logger,
+      coroutineName = "connectCoroutineScope@${logger.nameWithId}"
+    )
 
   suspend fun executeMutation(
     requestId: String,
@@ -361,6 +385,108 @@ internal class DataConnectGrpcRPCs(
     return cachedData?.let(ExecuteQueryResult::FromCache)
   }
 
+  suspend fun connect(
+    streamId: String,
+    callerSdkType: FirebaseDataConnect.CallerSdkType,
+    authToken: DataConnectAuth.GetAuthTokenResult?,
+    appCheckToken: DataConnectAppCheck.GetAppCheckTokenResult?,
+  ): DataConnectBidiConnectStream {
+    val metadata = grpcMetadata.get(authToken, appCheckToken, callerSdkType)
+    val kotlinMethodName = "connect()"
+
+    val initRequest: StreamRequest =
+      StreamRequest.newBuilder().run {
+        setRequestId("init")
+        setName(connectorResourceName)
+        build()
+      }
+
+    val outgoingRequests = Channel<StreamRequest>(capacity = UNLIMITED)
+
+    outgoingRequests.trySend(initRequest).let {
+      check(it.isSuccess) {
+        "internal error b2bs3s6n3c: outgoingRequests.trySend(initRequest) should have succeeded " +
+          "since the outgoingRequests Channel has capacity=UNLIMITED, but it did not: $it"
+      }
+    }
+
+    val outgoingRequestsFlow: Flow<StreamRequest> =
+      outgoingRequests.consumeAsFlow().onEach {
+        if (it === initRequest) {
+          logger.logGrpcSending(
+            requestId = streamId,
+            kotlinMethodName = kotlinMethodName,
+            grpcMethod = ConnectorStreamServiceGrpc.getConnectMethod(),
+            metadata = metadata,
+            request = { initRequest.toStructProto(authUid = authToken?.authUid) },
+            requestTypeName = "StreamRequest",
+            authUid = authToken?.authUid,
+          )
+        } else {
+          logger.logGrpcSending(
+            streamId = streamId,
+            requestId = it.requestId,
+            kotlinMethodName = kotlinMethodName,
+            request = { it.toStructProto(authUid = authToken?.authUid) },
+            requestTypeName = "StreamRequest",
+          )
+        }
+      }
+
+    val result = lazyStreamingGrpcStub.get().runCatching { connect(outgoingRequestsFlow, metadata) }
+
+    result.onFailure {
+      outgoingRequests.close(it)
+      logger.logGrpcFailed(
+        requestId = streamId,
+        kotlinMethodName = kotlinMethodName,
+        it,
+      )
+    }
+
+    val incomingResponses: Flow<StreamResponse> =
+      result
+        .getOrThrow()
+        .onEach {
+          logger.logGrpcReceived(
+            streamId = streamId,
+            requestId = it.requestId,
+            kotlinMethodName = kotlinMethodName,
+            response = { it.toStructProto() },
+            responseTypeName = "StreamResponse",
+          )
+        }
+        .onCompletion { throwable ->
+          outgoingRequests.close(
+            when (throwable) {
+              null -> CancellationException("stream $streamId completed")
+              is CancellationException -> throwable
+              else -> CancellationException("stream $streamId failed", throwable)
+            }
+          )
+
+          if (throwable === null) {
+            logger.logGrpcCompleted(
+              requestId = streamId,
+              kotlinMethodName = kotlinMethodName,
+            )
+          } else {
+            logger.logGrpcFailed(
+              requestId = streamId,
+              kotlinMethodName = kotlinMethodName,
+              throwable = throwable,
+            )
+          }
+        }
+
+    return DataConnectBidiConnectStream(
+      outgoingRequests,
+      incomingResponses,
+      connectCoroutineScope,
+      Logger("${logger.nameWithId} $kotlinMethodName [sid=$streamId]"),
+    )
+  }
+
   suspend fun getEmulatorInfo(requestId: String): EmulatorInfo {
     val request = GetEmulatorInfoRequest.getDefaultInstance()
     val kotlinMethodName = "getEmulatorInfo()"
@@ -433,6 +559,7 @@ internal class DataConnectGrpcRPCs(
   suspend fun close() {
     logger.debug { "close()" }
     mutex.withLock { closed = true }
+    connectCoroutineScope.cancel("DataConnectGrpcRPCs.close() called [xn8dqn8dzm]")
 
     val grpcChannel = lazyGrpcChannel.initializedValueOrNull
     val cacheDb = lazyCacheDb.initializedValueOrNull?.ref
@@ -463,6 +590,8 @@ internal class DataConnectGrpcRPCs(
     if (exceptions.isNotEmpty()) {
       throw exceptions.first().apply { exceptions.drop(1).forEach { addSuppressed(it) } }
     }
+
+    connectCoroutineScope.coroutineContext.job.join()
   }
 
   companion object {
@@ -533,6 +662,28 @@ internal class DataConnectGrpcRPCs(
         }
       }
       "$kotlinMethodName [rid=$requestId] sending: ${struct.toCompactString(keySortSelector)}"
+    }
+
+    private inline fun Logger.logGrpcSending(
+      streamId: String,
+      requestId: String,
+      kotlinMethodName: String,
+      request: () -> Struct,
+      requestTypeName: String,
+    ) = debug {
+      "$kotlinMethodName [sid=$streamId, rid=$requestId] sending $requestTypeName: " +
+        request().toCompactString()
+    }
+
+    private inline fun Logger.logGrpcReceived(
+      streamId: String,
+      requestId: String,
+      kotlinMethodName: String,
+      response: () -> Struct,
+      responseTypeName: String,
+    ) = debug {
+      "$kotlinMethodName [sid=$streamId, rid=$requestId] received $responseTypeName: " +
+        response().toCompactString()
     }
 
     private fun Logger.logGrpcReturningFromCache(